cli changed to path_helper because of no need to cli
