+++ config.py
@@ -1,0 +1,7 @@
+"""
+for project_path -> /imansh/pip-helper/
+for project_virtualenv_path -> path of virtual environment that you have set for your project
+"""
+
+project_path = '/Coding/GitHub/pip-helper/pip-helper'
+project_virtualenv_path = ''
